pyproject.toml
[project]
name = "maestro"
version = "0.1.0"
description = "AI Agent Orchestrator for coordinating multiple coding agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.1",
    "fastmcp>=2.14.5",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.3.2",
    "typer>=0.21.1",
    "uvicorn>=0.40.0",
]

[project.scripts]
maestro = "maestro.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["maestro*"]

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "pyrefly>=0.51.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "9.0"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    # Ignore resource warnings from aiosqlite's internal socket handling during cleanup
    "ignore:Exception ignored in.*socket.socket:pytest.PytestUnraisableExceptionWarning",
    # Ignore unclosed event loop warnings that occur during pytest cleanup
    "ignore:Exception ignored in.*BaseEventLoop:pytest.PytestUnraisableExceptionWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["maestro"]
branch = true
parallel = true
omit = [
    "**/__pycache__/*",
    "**/tests/*",
    "**/.venv/*",
]

[tool.coverage.paths]
source = ["maestro"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
fail_under = 80
show_missing = true
skip_covered = false
precision = 2

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

# Ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 88
fix = true
exclude = [
    ".claude",
    ".venv",
    "__pycache__",
    ".git",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
    "ASYNC",  # flake8-async
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # `zip()` without an explicit `strict=` parameter
    "ARG001", # unused function argument (often needed for interfaces)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG001",  # Unused function argument (fixtures)
    "ARG002",  # Unused method argument
    "S101",    # Use of assert
]

[tool.ruff.lint.isort]
known-first-party = ["maestro"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
